execution-core/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
ccxt==4.3.24
websockets==12.0
aioredis==2.0.1
python-multipart==0.0.9
